+++ src/de/qtc/rmg/internal/RMGOption.java
@@ -102,7 +102,9 @@ public enum RMGOption {
     THREADS("--threads", "maximum number of threads (default: 5)", Arguments.store(), RMGOptionGroup.ACTION, "threads"),
     YSO("--yso", "location of ysoserial.jar for deserialization attacks", Arguments.store(), RMGOptionGroup.ACTION, "yso-path"),
     DGC_METHOD("--dgc-method", "method to use for dgc operations", Arguments.store(), RMGOptionGroup.ACTION, "method"),
-    REG_METHOD("--registry-method", "method to use for registry operations", Arguments.store(), RMGOptionGroup.ACTION, "method");
+    REG_METHOD("--registry-method", "method to use for registry operations", Arguments.store(), RMGOptionGroup.ACTION, "method"),
+    SERIAL_VERSION_UID("--serial-version-uid", "serialVersionUID to use for RMI stubs", Arguments.store(), RMGOptionGroup.ACTION, "uid"),
+    PAYLOAD_SERIAL_VERSION_UID("--payload-serial-version-uid", "serialVersionUID to use for payload classes", Arguments.store(), RMGOptionGroup.ACTION, "uid");
 
 
     public final String name;
@@ -119,6 +121,7 @@ public enum RMGOption {
             RMGOption.CONN_FOLLOW, RMGOption.CONN_SSL, RMGOption.SSRF_GOPHER, RMGOption.SSRF, RMGOption.BIND_BYPASS, RMGOption.GUESS_CREATE_SAMPLES,
             RMGOption.GUESS_TRUSTED, RMGOption.GUESS_FORCE_GUESSING, RMGOption.GUESS_DUPLICATES, RMGOption.GUESS_UPDATE, RMGOption.GUESS_ZERO_ARG,
             RMGOption.ENUM_BYPASS, RMGOption.NO_CANARY, RMGOption.NO_PROGRESS, RMGOption.SSRF_ENCODE, RMGOption.SSRF_RAW);
+    private final static EnumSet<RMGOption> longOptions = EnumSet.of(RMGOption.SERIAL_VERSION_UID, RMGOption.PAYLOAD_SERIAL_VERSION_UID);
 
     /**
      * Initializes an enum field with the corresponding Option name, the Option description the argument action,
@@ -273,6 +276,9 @@ public enum RMGOption {
                     if( intOptions.contains(option) )
                         defaultValue = Integer.valueOf((String) defaultValue);
 
+                    else if( longOptions.contains(option) )
+                        defaultValue = Long.valueOf((String) defaultValue);
+
                     else if( booleanOptions.contains(option) )
                         defaultValue = Boolean.valueOf((String) defaultValue);
 
@@ -355,8 +361,12 @@ public enum RMGOption {
         } else if( option == RMGOption.DGC_METHOD ) {
             arg.choices("clean", "dirty");
 
-        } else if( intOptions.contains(option) )
+        } else if( intOptions.contains(option) ) {
             arg.type(Integer.class);
+
+        } else if( longOptions.contains(option) ) {
+            arg.type(Long.class);
+        }
     }
 
     /**