+++ docker/example-server/CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [4.0] - Jan 19, 2023
+
+### Added
+
+* Add build script for containers
+* Add build script for certificate store
+
+### Changed
+
+* Change RMI hostname from `iinsecure.dev` to `iinsecure.example`
+* Change server certificate
+
+
 ## [3.3] - May 08, 2022
 
 ### Changed