+++ docker/ssrf-server/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.4] - Jan 19, 2023
+
+### Changed
+
+* Remove unused hostname from startup script
+
+
 ## [1.3] - May 08, 2022
 
 ### Changed